LPS-153402 Move DDMExpressionBuilder related code to Util class
